pyproject.toml
[project]
name = "ruchky-backend"
version = "0.1.0"
description = "Na Ruchky DRF backend and Admin panel"
readme = "README.md"
requires-python = ">=3.13.1"
dependencies = [
    "django>=5.1.4",
    "django-allauth>=65.3.1",
    "psycopg[binary,pool]>=3.2.3",
    "python-dotenv>=1.0.1",
]
